fix(programmetype): use previous groups programmetypeid for new groups EC-93

+++ src/api/Libraries/Modules/ECDLink.DataAccessLayer/Entities/IncomeStatements/StatementsIncome.cs
@@ -19,9 +19,9 @@ namespace ECDLink.DataAccessLayer.Entities.IncomeStatements
         public string? ContributionTypeId { get; set; }
         public string? PayTypeId { get; set; }
         public string? IncomeTypeId { get; set; }
-        public decimal Amount { get; set; }
-        public decimal AmountExpected { get; set; }
-        public decimal ChildCoverAmount { get; set; }        
+        public double Amount { get; set; }
+        public double AmountExpected { get; set; }
+        public double ChildCoverAmount { get; set; }        
         public string Description { get; set; }
         public string PhotoProof { get; set; }
         public bool Submitted { get; set; }
